pyproject.toml
[project]
name = "whatsonthisdomain"
version = "0.1.0"
description = "Recon and attack surface monitoring pipeline for bug bounty and authorized pentesting"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.15",
    "rich>=13",
    "sqlalchemy[asyncio]>=2",
    "aiosqlite>=0.20",
    "alembic>=1.13",
    "pyyaml>=6",
    "python-dotenv>=1",
    "httpx>=0.27",
    "aiodns>=3",
    "pydantic>=2",
    "platformdirs>=4",
]

[project.scripts]
wotd = "wotd.cli:app"
whatsonthisdomain = "wotd.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["wotd"]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "pre-commit>=4",
    "mypy>=1.11",
    "types-pyyaml>=6",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "UP",   # pyupgrade
    "RUF",  # ruff-specific
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
